+++ src/components/ViewPager/DefaultViewPageIndicator.js
